avoid negative left shifts
The literal "0" is treated as a signed int, which results in the expression
"~0 << ..." performing a left shift of a negative number. This is undefined
behaviour in C which newer GCC and Clang will warn about with
-Wshift-negative-value. This commit uses unsigned long literals to ensure the
shifts are interpreted correctly. For the record, all the compilers I have on
hand handle this code correctly, but other compilers may not.